server/requirements.txt
boto3==1.16.59
django==3.1.5
django-cors-headers==3.6.0
django-filter==2.4.0
django-storages==1.11.1
djangorestframework==3.12.2
Pillow==8.1.0
